feat: move api gateway functionality to sls-aws
