+++ packages/sls-aws/src/api-gateway/index.ts
@@ -1,4 +1,5 @@
 export { createSuccessResponse } from './response'
 export { apiGatewayV2 } from './handler'
 export * from './payload'
+export * from './dependency'
 export type { APIGatewayV2Handler } from './handler'
