+++ packages/sls-env/package.json
@@ -20,7 +20,6 @@
     "@types/pino": "^6.3.5"
   },
   "devDependencies": {
-    "@types/aws-lambda": "^8.10.59",
     "pino": "^6.11.0"
   }
 }
